\name{FeistyR-package}
\alias{FeistyR-package}
\alias{FeistyR}
\docType{package}
\title{
  The Feisty model: FishErIes Size and functional TYpe. 
}
\description{
Implements the Feisty model as described in Van Denderen et al 2020 and Petrik et al., 2019
     }
\details{
  \tabular{ll}{
    Package: \tab FeistyR\cr
    Type: \tab Package\cr
    Created: \tab 2022-31-01\cr
    License: \tab GNU Public License 2 or above\cr
  }
}
\author{
  Ken H Andersen
  Karline Soetaert 
}
\seealso{
\itemize{
 \item \code{\link{simulateFeisty}} Dynamic run of the feisty model
 \item \code{\link{setupPelagicSpecies}} Parameter input of the feisty model
}
}
\examples{
\dontrun{
## show examples (see respective help pages for details)
example(simulateFeisty)
}
}

\references{
Van Denderen et al. 2020. Emergent global biogeography of marine fish food webs. 
Global Ecology and Biogeography, DOI: 10.1111/geb.13348 

Petrik, CM, Stock, CA, Andersen, KH, van Denderen, PD, Watson, JR 2019. 
Bottom-up drivers of global patterns of demersal, forage, and pelagic fishes. 
Progress in Oceanography, 176, 102124. DOI: 10.1016/j.pocean.2019.102124
}
\keyword{ package }
