\name{Interactive Tool for Feisty}
\alias{webFeisty}
\title{
  Interactive visualisation of the FishErIes Size and functional TYpe model.
}
\description{
  A shiny interface to simulate fish dynamics.
}

\author{
  Karline Soetaert, Ken H Andersen 
}

\usage{
 webFeisty(...)
}

\arguments{
  \item{... }{Arguments passed to the shiny application (not used).
  }
}

\examples{
\dontrun{
  webFeisty()
}  
}

\keyword{ manip }
